// app/src/generated-api/models/ExitRoom.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Underworld
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.5.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ExitRoom
 */
export interface ExitRoom {
    /**
     * 
     * @type {string}
     * @memberof ExitRoom
     */
    exit_id: string;
}

export function ExitRoomFromJSON(json: any): ExitRoom {
    return ExitRoomFromJSONTyped(json, false);
}

export function ExitRoomFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExitRoom {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'exit_id': json['exit_id'],
    };
}

export function ExitRoomToJSON(value?: ExitRoom | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'exit_id': value.exit_id,
    };
}